Code split
Detekt function size limit extended to 120 lines
