+++ app/src/main/java/com/wndenis/snipsnap/utils/Utils.kt
@@ -1,10 +1,8 @@
-package com.wndenis.snipsnap
+package com.wndenis.snipsnap.utils
 
 import android.app.Activity
 import android.content.Context
 import android.view.inputmethod.InputMethodManager
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
 
 const val MIN_FILE_LENGTH = 5
 
@@ -13,11 +11,6 @@ const val DAY_SEC = HOUR_SEC * 24
 const val WEEK_SEC = DAY_SEC * 7
 const val MONTH_SEC = DAY_SEC * 30 // wndenis: rough approx
 const val YEAR_SEC = DAY_SEC * 365
-
-fun LocalDateTime.conv(): String {
-    val df = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm")
-    return this.format(df)
-}
 
 fun hideKeyboard(context: Context) {
     val imm = context.getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
